Factor a NewHeadHandler from code in Node, add integration tests for message handling. (#3265)
Extract fake chain provider, validator, and mock publisher for re-use.
Factor a NewHeadHandler from code in Node, add integration tests for message handling.

This allows demonstration of the bug reported in #3052.

+++ core/policy.go
@@ -48,8 +48,8 @@ type DefaultQueuePolicy struct {
 
 // NewMessageQueuePolicy returns a new policy which removes mined messages from the queue and expires
 // messages older than `maxAgeTipsets` rounds.
-func NewMessageQueuePolicy(messages MessageProvider, maxAge uint64) *DefaultQueuePolicy {
-	return &DefaultQueuePolicy{messages, maxAge}
+func NewMessageQueuePolicy(messages MessageProvider, maxAge uint) *DefaultQueuePolicy {
+	return &DefaultQueuePolicy{messages, uint64(maxAge)}
 }
 
 // HandleNewHead removes from the queue all messages that have now been mined in new blocks.
